+++ tools/webTools/lib/src/util.js
@@ -411,7 +411,7 @@ $("#module-ok-askdeleteTag").click(() => {
 $("#module-ok-postNotice").bind("click", function() {
     const form = $(this).parent().parent().find("form");
     const notice = form.find("textarea").val();
-    websocket.send(`node-manager->tagsService->createTag(${JSON.stringify(data)})`);
+    websocket.send(`node-manager->systemService->postSystemNotice("${notice}")`);
 });
 
 // 发布公告
@@ -419,6 +419,5 @@ $("#module-ok-postAnnouncements").bind("click", function() {
     const form = $(this).parent().parent().find("form");
     const title = form.find("input").val();
     const content = form.find("textarea").val();
-
-
+    websocket.send(`node-manager->systemService->postAnnouncement("${title}" | "${content}")`);
 });